add LoadingSpinner and move NavigationItems
