+++ src/pages/auth/new-user.tsx
@@ -1,3 +1,4 @@
+import LoadingSpinner from "@components/ui/LoadingSpinner";
 import { getServerAuthSession } from "@server/auth";
 import { api } from "@utils/api";
 import { type GetServerSidePropsContext } from "next";
@@ -90,30 +91,7 @@ const NewUser = () => {
         aria-disabled={updateUser.isLoading}
         className="flex h-10 w-32 items-center justify-center rounded-full bg-green-500 py-2 px-8 font-bold text-white hover:bg-green-600"
       >
-        {updateUser.isLoading ? (
-          <svg
-            className="h-4 w-4 animate-spin text-white"
-            xmlns="http://www.w3.org/2000/svg"
-            fill="none"
-            viewBox="0 0 24 24"
-          >
-            <circle
-              className="opacity-25"
-              cx="12"
-              cy="12"
-              r="10"
-              stroke="currentColor"
-              strokeWidth="4"
-            ></circle>
-            <path
-              className="opacity-75"
-              fill="currentColor"
-              d="M4 12a8 8 0 018-8V0C5.373 0 0 5.373 0 12h4zm2 5.291A7.962 7.962 0 014 12H0c0 3.042 1.135 5.824 3 7.938l3-2.647z"
-            ></path>
-          </svg>
-        ) : (
-          <span>Confirm</span>
-        )}
+        {updateUser.isLoading ? <LoadingSpinner /> : <span>Confirm</span>}
       </button>
     </form>
   );
